Include service provider API format when returning transformed params (#5)
* Add openrouter as a service provider to the default config and use openrouter to access Gemma 3 27b in the default example config

* Expose service provider API format information to client applications, enabling them to select appropriate client classes. e.g. xAI has `api_format: openai`.
Add openrouter as a service provider to the default config and use it for Gemma 3. OpenRouter is used for Gemma 3 largely due to the system prompt issue for Gemma when called via `google-genai` - Gemma does not accept a developer/system prompt. Therefore, calling it via the OpenAI format with OpenRouter works better, and Gemma is free, so users will not incur any markup. Alternatively, FireworksAI serves Gemma 3 at a low $0.10 /mtok cost.

* Add tests

* Remove deprecated models and update Gemini model configurations

* Update docs and examples to include the new `api_format` returned from the parameter transformer